+++ admin/page/blog/post.php
@@ -8,10 +8,14 @@
 
         $file_name = $_FILES["file"]["name"];
         $tmp_name = $_FILES["file"]["tmp_name"];
-        move_uploaded_file($tmp_name, "../images/blog/".$file_name);
 
-        mysqli_query($connection, "INSERT INTO post VALUES('','$category_id','$title','$name','$description','$file_name','$date','0')");
-        header("location:index.php?post");
+        if($file_name!=="" || !empty($file_name) && !empty($category_id) && !empty($title) && !empty($name) && !empty($description)){
+            move_uploaded_file($tmp_name, "../images/blog/".$file_name);
+            mysqli_query($connection, "INSERT INTO post VALUES('','$category_id','$title','$name','$description','$file_name','$date','0')");
+            header("location:index.php?post&success#post-success");
+        } else {
+            header("location:index.php?post&failed#post-failed");
+        }
     }
 
     $post = mysqli_query($connection, "SELECT post.*, category.category_name FROM post, category
@@ -54,7 +58,7 @@
                                 <label>Author</label>
                                 <input class="form-control" type="text" name="name" />
                             </div>
-                            <div class="form-group">
+                            <div class="form-group" id="post-failed">
                                 <label>Description</label>
                                 <textarea class="form-control" rows="3" name="description" id="editor1"></textarea>
                             </div>
@@ -62,6 +66,15 @@
                                 <label>Image</label>
                                 <input type="file" name="file" />
                             </div>
+                            <?php if(isset($_GET["failed"])) {?>
+                                <div class="alert alert-danger" role="alert">
+                                    Lengkapi Data Post!
+                                </div>
+                            <?php } else if(isset($_GET["success"])) { ?>
+                                <div class="alert alert-success" role="alert">
+                                    Data Post Berhasil Dipublish
+                                </div>
+                            <?php } ?>
                             <button type="submit" name="submit" class="btn btn-success">Save</button>
                             <button type="reset" class="btn btn-warning">Reset</button>
                         </form>
@@ -73,7 +86,7 @@
             <div class="panel-heading">
                 List Data
             </div>
-            <div class="panel-body">
+            <div class="panel-body" id="post-success">
                 <div class="table-responsive">
                     <!-- <table class="table table-striped table-bordered table-hover" id="dataTables-example"> -->
                     <table id="table_admin" class="display">
@@ -84,6 +97,7 @@
                                 <th>Title</th>
                                 <th>Author</th>
                                 <th>Description</th>
+                                <th>Publish</th>
                                 <th>View</th>
                                 <th>Image</th>
                                 <th>Update</th>
@@ -100,10 +114,11 @@
                                         <td><?php echo $row_post["title"] ?></td>
                                         <td><?php echo $row_post["name"] ?></td>
                                         <td><?php echo substr($row_post["description"],0,200)."..." ?></td>
+                                        <td><?php echo tanggal_indonesia($row_post["date"]) ?></td>
                                         <td><?php echo $row_post["view"] ?></td>
                                         <td><img src="../images/blog/<?php echo $row_post["image"] ?>" width="88" class="img-responsive" /></td>
                                         <td class="center"><a href="index.php?post-update=<?php echo $row_post["id"] ?>" class="btn btn-primary btn-xs" type="button">Update</a></td>
-                                        <td class="center"><a href="index.php?post-delete=<?php echo $row_post["id"] ?>" class="btn btn-primary btn-xs" type="button">Delete</a></td>
+                                        <td class="center"><a href="index.php?post-delete=<?php echo $row_post["id"] ?>" class="btn btn-primary btn-xs" type="button" onclick="return confirm('Delete Data Post?')">Delete</a></td>
                                     </tr>
                                 <?php $no++; } ?>
                             <?php } ?>
